bugfix in shift_* methods
* It is an error to shift a node with its subtree after/before one of its own descendants.
Such operation is not well defined and almost surely means an error in the calling code.
* On the other hand, it is quite common to shift node A (with or without its subtree)
after/before its parent (or higher ancerstor) B's subtree.
In this case, we need to exclude subtree A from the subtree B's span
because of non-projectivities. E.g.

node | parent
-------------
B    | root
X    | root
A    | B

`A.shift_after_subtree(B)` should result in ordering "B A X",
not "B X A" (as it was in Pytreex until now).

In Python I could not use `descendats(except=node)`
because `except` is a keyword, so I called it `except_subtree`.
Maybe it could be renamed to `exclude`.