another bugfix in shift_* methods
`nodeA.shift_node_after(nodeA)` should result in no action
(maybe a warning), but in Pytreex it resulted in changes in word order.

`nodeA.shift_subtree_after(nodeA)` should also result in no action.
Here, one could imagine that this way you'll make the subtree of `nodeA`
magically projective, but this is not a good idea.
First, it is not clear where to shift the nodes causing
the non-projectivities, ie. non-descendants of `nodeA` which are in its span
(whether before or after the subtree of `nodeA`).
Second, such behavior would be cryptic and it is better to have a special method
for such projectivization.
Third, it would not cancel non-projectivities "inside" the subtree of `nodeA`,
so I don't see any use case for such semi-projectivization.